# Conflicts:
#	ふぉるてっしも！.vcxproj.filters

+++ Library/DebugProcess.cpp
@@ -76,7 +76,7 @@ void UpdateDebugProcess(void)
 //=============================================================================
 void DrawDebugProcess(void)
 {
-	RECT rect = {0, 0, SCREEN_WIDTH, SCREEN_HEIGHT};
+	RECT rect = {0, 0, (long)SCREEN_WIDTH, (long)SCREEN_HEIGHT};
 
 	// î•ñ•\Ž¦
 	Dx9DebugFont->DrawText(NULL, DebugString, -1, &rect, DT_LEFT, D3DCOLOR_ARGB(0xFF, 0x00, 0xFF, 0xFF));
